Split optimizer into modules
Some parts might be replaced, or alternatives evaluated; this makes it
easier to add per-module tests and track the flow of data.
